Fix CR comments.
