+++ lib/assets/javascripts/cartodb3/components/modals/dataset-metadata/dataset-metadata-view.js
@@ -1,6 +1,6 @@
 var Backbone = require('backbone');
 var CoreView = require('backbone/core-view');
-var VisMetadataModel = require('./dataset-metadata-model');
+var VisMetadataModel = require('../../../data/vis-metadata-model');
 var template = require('./dataset-metadata.tpl');
 var FooterView = require('./footer/footer-view');
 var templateError = require('./dataset-metadata-error.tpl');
@@ -26,8 +26,13 @@ module.exports = CoreView.extend({
     this._visDefinitionModel = opts.visDefinitionModel;
     this._configModel = opts.configModel;
 
-    this._visMetadataModel = new VisMetadataModel({}, {
-      visDefinitionModel: this._visDefinitionModel
+    this._visMetadataModel = new VisMetadataModel({
+      name: this._visDefinitionModel.get('name'),
+      description: this._visDefinitionModel.get('description'),
+      license: this._visDefinitionModel.get('license'),
+      source: this._visDefinitionModel.get('source'),
+      attributions: this._visDefinitionModel.get('attributions'),
+      tags: this._visDefinitionModel.get('tags')
     });
 
     this._stateModel = new Backbone.Model({
@@ -45,13 +50,13 @@ module.exports = CoreView.extend({
     } else if (this._isLoading()) {
       this._renderLoading();
     } else {
+      this.$el.html(template());
       this._initViews();
     }
+    return this;
   },
 
   _initViews: function () {
-    this.$el.html(template());
-
     var formView = new FormView({
       visDefinitionModel: this._visDefinitionModel,
       visMetadataModel: this._visMetadataModel,
