+++ lib/assets/javascripts/cartodb3/components/modals/dataset-metadata/form/form-view.js
@@ -23,13 +23,12 @@ module.exports = CoreView.extend({
 
   render: function () {
     this.clearSubViews();
+    this.$el.html(templateForm());
     this._initViews();
     return this;
   },
 
   _initViews: function () {
-    this.$el.html(templateForm());
-
     this._addNameView();
     this._addDescriptionView();
     this._addSourceView();
