+++ lib/assets/javascripts/cartodb3/components/modals/map-metadata/form/form-view.js
@@ -22,13 +22,12 @@ module.exports = CoreView.extend({
 
   render: function () {
     this.clearSubViews();
+    this.$el.html(templateForm());
     this._initViews();
     return this;
   },
 
   _initViews: function () {
-    this.$el.html(templateForm());
-
     this._addInputView();
     this._addTextareaView();
     this._addTagsView();
