+++ lib/assets/javascripts/cartodb3/components/modals/map-metadata/map-metadata-view.js
@@ -1,6 +1,6 @@
 var Backbone = require('backbone');
 var CoreView = require('backbone/core-view');
-var VisMetadataModel = require('./vis-metadata-model');
+var VisMetadataModel = require('../../../data/vis-metadata-model');
 var template = require('./map-metadata.tpl');
 var FooterView = require('./footer/footer-view');
 var templateError = require('./map-metadata-error.tpl');
@@ -27,8 +27,10 @@ module.exports = CoreView.extend({
     this._modalModel = opts.modalModel;
     this._visDefinitionModel = opts.visDefinitionModel;
 
-    this._visMetadataModel = new VisMetadataModel({}, {
-      visDefinitionModel: this._visDefinitionModel
+    this._visMetadataModel = new VisMetadataModel({
+      name: this._visDefinitionModel.get('name'),
+      description: this._visDefinitionModel.get('description'),
+      tags: this._visDefinitionModel.get('tags')
     });
 
     this._stateModel = new Backbone.Model({
@@ -46,13 +48,13 @@ module.exports = CoreView.extend({
     } else if (this._isLoading()) {
       this._renderLoading();
     } else {
+      this.$el.html(template());
       this._initViews();
     }
+    return this;
   },
 
   _initViews: function () {
-    this.$el.html(template());
-
     var formView = new FormView({
       visDefinitionModel: this._visDefinitionModel,
       visMetadataModel: this._visMetadataModel
